src/requirements.txt
aiofiles==0.6.0
celery==5.0.5
fastapi==0.64.0
flower==1.0.0
Jinja2==3.0.3
pytest==6.2.4
redis==3.5.3
requests==2.25.1
uvicorn==0.13.4
aioredis==1.3.1
pydantic==1.6.1
ujson==4.1.0
networkx==2.6.3
starlette-prometheus==0.8.0
